STYLE: split sprite.py in 2 files (motion.py and sprite.py)
